+++ rapidenv/osh/download.py
@@ -1,67 +1,54 @@
+import urllib.request
+import os
 import shutil
 from pathlib import Path
 
-def download(url, dst, makedirs=True, unpack=False, skip_dst_exists=False):
+
+def download_archive(url, dst, makedirs=True, skip_dst_exists=False, delete_archive=True):
     """
-    copy (file) or copytree (dir)
-    if src doesn't exist FileExistsError is thrown
+    download archive from url ro dst parent directory and unpack it to dst folder.
+    if archive contains single folders recursively remove them placing main folder in dst.
 
-    src        dst            action
-    ----       -------------  ---------------------------------------
-    url(file)  file           dst file is overwritten
-    url(file)  dir            src file is copied to dst/file
-    url(file)  doesn't exist  src file is copied to dst path
-
-    :param src: source url, source must be file url
-    :param dst: destination path
-    :param makedirs: create dst parent dir tree if not exists
-    :param unpack: unpack downloaded file (.zip or .tar)
-    :param skip_dst_exists: do nothing if dst exists
+    :param src source url, source must be and archive file url
+    :param dst destination directory
+    :param makedirs create dst parent dir tree if not exists
+    :param skip_dst_exists do nothing if dst exists
+    :param delete_archive delete archive downloaded once done
     :return:
     """
-
-    import urllib.request
-    import os
-
     dst = Path(dst)
-
-    # explicit dst path in case src is file and dst is dir
-    if dst.is_dir():
-        dst = dst / url.split('/')[-1]
 
     # do nothing if dst exists and skip_dst_exists flag is raised
     if skip_dst_exists and dst.exists():
         return
+
+    # validate destination folder doesn't already exists
+    if dst.exists():
+        raise Exception("Archive unpack destination already exists", str(dst))
 
     # create parent dir if doesn't exists
     if makedirs and dst.parent != dst and not dst.parent.exists():
         os.makedirs(dst.parent)
 
     # download file
-    urllib.request.urlretrieve(url, dst)
+    archive = dst.parent / url.split('/')[-1]
+    urllib.request.urlretrieve(url, archive)
 
-    if unpack:
-        s_dst = str(dst)
-        suffix = None
-        supported = ['.zip', '.tar.gz']
-        for s in supported:
-            if s_dst.endswith(s):
-                suffix = s
-        if suffix is None:
-            raise Exception(f"Unsupported archive type to unpack {dst}, supported: {', '.join(supported)}")
+    shutil.unpack_archive(archive, extract_dir=dst)
 
-        s_dst_trim = s_dst[:-len(suffix)]
+    # check if unpacked folder contains only 1 folder (recursively), if so move its content to parent folder
+    content = os.listdir(dst)
+    while len(content) == 1 and Path(content[0]).is_dir:
+        shutil.move(f"{dst}", f"{dst}_temp")
+        shutil.move(f"{dst}_temp/{content[0]}", dst)
+        os.rmdir(f"{dst}_temp")
+        content = os.listdir(dst)
 
-        shutil.unpack_archive(s_dst, extract_dir=s_dst_trim)
-
-        # check if unpacked folder contains only 1 folder, if so move its content to parent folder
-        content = os.listdir(s_dst_trim)
-        if len(content) == 1 and Path(content[0]).is_dir:
-            shutil.move(f"{s_dst_trim}", f"{s_dst_trim}_temp")
-            shutil.move(f"{s_dst_trim}_temp/{content[0]}", str(dst.parent / content[0]))
-            os.rmdir(f"{s_dst_trim}_temp")
+    if delete_archive:
+        os.remove(archive)
 
 
-if __name__ == __main__:
-    url = 'https://www.tcpdump.org/release/libpcap-1.10.0.tar.gz'
-    download(url, './tmp')
+if __name__ == "__main__":
+    # url = 'https://www.tcpdump.org/release/libpcap-1.10.0.tar.gz'
+    url = "https://file-examples-com.github.io/uploads/2017/02/zip_2MB.zip"
+    download_archive(url, './tmp/zip_2MB')