+++ tests/test_oslib.py
@@ -1,2 +1,0 @@
-# example for run process string containing spaces test input
-# run_process('pytest -m "not dev and not me" and "not by" and \'not prev\'', cwd="./tests")